Retrieve bus encryption flags from drive and content certificate
For bus encryption to be active it must be enabled both by the drive
and the disc.
Drive certificate contains Bus Encryption Capable (BEC) flag in the
least significant bit in the second byte of the certificate.
Content certificate contains Bus Encryption Enabled (BEE) flag in the
most significant bit in the second byte of the certificate.
Retrieve both bits so it is now possible to check if bus encryption is
active (it will not be active if the disc does not enable it, even on
bus encryption drive; most discs currently do not enable it).