Fix reading PMSN
While working on bus encryption I've found out that some drives (if
not all) require a new AACS-Auth for reading data keys. It seems that
the same goes for reading PMSN. After digging through my discs I've
finally found one with PMSN and I was able to verify this with two LG
drives. Fix it by moving PMSN reading to a separate function with a
new AACS-Auth.