Do not retrieve MKB using MMC commands during AACS-Auth
In _mmc_read_auth() MMC commands are used to retrieve MKB from the
drive to display a warning if a host certificate is
revoked. Unfortunately, reading MKB using MMC is very slow (up to 10
seconds) and it bypasses OS cache so it takes the same amount of time
every time. _mmc_read_auth() is called several times (VID retrieval,
PMSN retrieval, read data key retrieval), which can cause up to 30
second delay when trying to play a disc (VID is cached, but PMSN is
not and BD-J code tends to try to retrieve it even when there is no
PMSN on the disc).

Fix this by removing MKB retrieval in _mmc_read_auth(). This only
causes a warning not to be displayed, however there still will be an
information if the host certificate is revoked (drive will simply
reject it).

 1 file changed, 67 deletions(-)