mmc_device_darwin: IODestroyPlugInInterface plugin interface
Store a reference to the plugin interface in the mmc struct and call
IODestroyPlugInInterface instead of Destroy(), as it is explicitly
documented that IODestroyPlugInInterface should be used and not
Destroy.
As destroying the plugin interface might stop the IOBDService,
call it only when we are done with anything needing it.